+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ### [Unreleased]
 #### Fixed
-- Commands `opponent_moves`, [GoPlaceArmiesCommand](src/Command/GoPlaceArmiesCommand.php) and [GoAttackTransferCommand](src/Command/GoAttackTransferCommand.php) fixed.
+- Commands [GoPlaceArmiesCommand](src/Command/GoPlaceArmiesCommand.php) and [GoAttackTransferCommand](src/Command/GoAttackTransferCommand.php) fixed.
 - Fixed some command class names.
 - Fixed `composer.json`: set _type_ to `library`, added directory `tests` to section _autoload_.
 #### Changed
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - Interface [Move](src/Game/Move/Move) and move [PickMove](src/Game/Move/PickMove) added.
 - Added class [Round](src/Game/Round.php) to save map snapshots and player moves of each round.
 - Added class [Environment](src/Game/Environment.php) as container for player, map and rounds.
+- Added missing command [OpponentMovesCommand](src/Command/OpponentMovesCommand.php).
 - Added configuration file and `@covers` annotations for code coverage.
 - Added missing unit test for command classes.
 - Added [CHANGELOG.md](CHANGELOG.md) file.